requirements.txt
python-telegram-bot
fpdf
polars
scipy
numpy
requests
beautifulsoup4
